LayoutTests/storage/domstorage/localstorage/simple-usage-expected.txt

This test tries simple operations on localStorage
Length is 0
Value for FOO is null
Length is 1
Value for FOO is BAR
Key for index 0 is FOO
Correctly caught exception trying to access key at index 1
Error: INDEX_SIZE_ERR: DOM Exception 1
Length is 1
Value for FOO is BAZ
Length is 0
Value for FOO is null
